更新了几个syscall

+++ typedef.h
@@ -5,7 +5,7 @@ typedef unsigned char byte;
 typedef unsigned short zjie;
 typedef unsigned short zchr;
 typedef unsigned int word;
-typedef unsigned long dwrd;
+typedef unsigned long long dwrd;
 typedef dwrd dword;
 
 #endif
